Integrate with wrapper library
