+++ src/configuration.js
@@ -1,0 +1,84 @@
+export const configuration = {
+  stimuli: {
+    'alien1_a1.png': 'images/alien1_a1.png',
+    'alien4_sp.png': 'images/alien4_sp.png',
+    'tutalien1_a2.png': 'images/tutalien1_a2.png',
+    'tutalien4_deact.png': 'images/tutalien4_deact.png',
+    'alien1_a2.png': 'images/alien1_a2.png',
+    'tutalien1_deact.png': 'images/tutalien1_deact.png',
+    'tutalien4_norm.png': 'images/tutalien4_norm.png',
+    'alien1_deact.png': 'images/alien1_deact.png',
+    'earth.jpg': 'images/earth.jpg',
+    'tutalien1_norm.png': 'images/tutalien1_norm.png',
+    'tutalien4_sp.png': 'images/tutalien4_sp.png',
+    'alien1_norm.png': 'images/alien1_norm.png',
+    'nothing.png': 'images/nothing.png',
+    'tutalien1_sp.png': 'images/tutalien1_sp.png',
+    'tutgreenplanet.jpg': 'images/tutgreenplanet.jpg',
+    'alien1_sp.png': 'images/alien1_sp.png',
+    'purpleplanet.jpg': 'images/purpleplanet.jpg',
+    'alien2_a1.png': 'images/alien2_a1.png',
+    'redplanet1.jpg': 'images/redplanet1.jpg',
+    'tutalien2_a1.png': 'images/tutalien2_a1.png',
+    'tutrocket1_a1.png': 'images/tutrocket1_a1.png',
+    'alien2_a2.png': 'images/alien2_a2.png',
+    'rocket1_a1.png': 'images/rocket1_a1.png',
+    'tutalien2_a2.png': 'images/tutalien2_a2.png',
+    'tutrocket1_a2.png': 'images/tutrocket1_a2.png',
+    'alien2_deact.png': 'images/alien2_deact.png',
+    'rocket1_a2.png': 'images/rocket1_a2.png',
+    'tutalien2_deact.png': 'images/tutalien2_deact.png',
+    'tutrocket1_deact.png': 'images/tutrocket1_deact.png',
+    'alien2_norm.png': 'images/alien2_norm.png',
+    'rocket1_deact.png': 'images/rocket1_deact.png',
+    'tutalien2_norm.png': 'images/tutalien2_norm.png',
+    'tutrocket1_norm.png': 'images/tutrocket1_norm.png',
+    'alien2_sp.png': 'images/alien2_sp.png',
+    'rocket1_norm.png': 'images/rocket1_norm.png',
+    'tutalien2_sp.png': 'images/tutalien2_sp.png',
+    'tutrocket1_sp.png': 'images/tutrocket1_sp.png',
+    'alien3_a1.png': 'images/alien3_a1.png',
+    'rocket1_sp.png': 'images/rocket1_sp.png',
+    'alien3_a2.png': 'images/alien3_a2.png',
+    'rocket2_a1.png': 'images/rocket2_a1.png',
+    'tutalien3_a1.png': 'images/tutalien3_a1.png',
+    'tutrocket2_a1.png': 'images/tutrocket2_a1.png',
+    'alien3_deact.png': 'images/alien3_deact.png',
+    'rocket2_a2.png': 'images/rocket2_a2.png',
+    'tutalien3_a2.png': 'images/tutalien3_a2.png',
+    'tutrocket2_a2.png': 'images/tutrocket2_a2.png',
+    'alien3_norm.png': 'images/alien3_norm.png',
+    'rocket2_deact.png': 'images/rocket2_deact.png',
+    'tutalien3_deact.png': 'images/tutalien3_deact.png',
+    'tutrocket2_deact.png': 'images/tutrocket2_deact.png',
+    'alien3_sp.png': 'images/alien3_sp.png',
+    'rocket2_norm.png': 'images/rocket2_norm.png',
+    'tutalien3_norm.png': 'images/tutalien3_norm.png',
+    'tutrocket2_norm.png': 'images/tutrocket2_norm.png',
+    'alien4_a1.png': 'images/alien4_a1.png',
+    'rocket2_sp.png': 'images/rocket2_sp.png',
+    'tutalien3_sp.png': 'images/tutalien3_sp.png',
+    'tutrocket2_sp.png': 'images/tutrocket2_sp.png',
+    'alien4_a2.png': 'images/alien4_a2.png',
+    't.png': 'images/t.png',
+    'tutyellowplanet.jpg': 'images/tutyellowplanet.jpg',
+    'alien4_deact.png': 'images/alien4_deact.png',
+    'tutalien4_a1.png': 'images/tutalien4_a1.png',
+    'alien4_norm.png': 'images/alien4_norm.png',
+    'tutalien1_a1.png': 'images/tutalien1_a1.png',
+    'tutalien4_a2.png': 'images/tutalien4_a2.png',
+    'fish.png': 'images/fish.png',
+    'tiger.png': 'images/tiger.png',
+    'shark.png': 'images/shark.png',
+    'turtle.png': 'images/turtle.png',
+    'replay.png': 'images/replay.png',
+    'earth.jpg': 'images/earth.jpg',
+    'tutgreenplanet.jpg': 'images/tutgreenplanet.jpg',
+    'tutyellowplanet.jpg': 'images/tutyellowplanet.jpg',
+    'blackbackground.jpg': 'images/blackbackground.jpg',
+    'tutgreenplanet.jpg': 'images/tutgreenplanet.jpg',
+    'earth.jpg': 'images/earth.jpg',
+    'tutyellowplanet.jpg': 'images/tutyellowplanet.jpg',
+    'earth.jpg': 'images/earth.jpg',
+  },
+};
