+++ src/css/styles.css
@@ -8,6 +8,8 @@
 .jspsych-content-wrapper {
   margin: auto;
   width: 100%;
+  display: block;
+  flex: none;
 }
 
 .jspsych-content {
