+++ src/lib/instructions.js
@@ -1,17 +1,3 @@
-// General jsPsych imports
-import 'jspsych';
-import 'jspsych/plugins/jspsych-audio-button-response';
-import 'jspsych/plugins/jspsych-audio-keyboard-response';
-import 'jspsych/plugins/jspsych-call-function';
-import 'jspsych/plugins/jspsych-fullscreen';
-import 'jspsych/plugins/jspsych-html-button-response';
-import 'jspsych/plugins/jspsych-html-keyboard-response';
-import 'jspsych/plugins/jspsych-image-keyboard-response';
-import 'jspsych/plugins/jspsych-instructions';
-import 'jspsych/plugins/jspsych-survey-likert';
-import 'jspsych/plugins/jspsych-survey-multi-choice';
-import 'jspsych/plugins/jspsych-survey-text';
-
 // Styling
 export const block_trials = 50; // total number of trials in each block
 export const transprob = 0.7; // probability of 'correct' transition
@@ -42,9 +28,6 @@ export const instructions_payoff = ['.9', '.1', '.9', '.1'];
 
 export const red_planet_first_rocket = (Math.random() < 0.5);
 
-export const reward_string = 'images/t.png';
-export const null_string = 'images/nothing.png';
-
 export const practice_pressing_idx = 5;
 export const practice_pressing_num = 4; // 4 trials to practice selecting alien
 export const practice_reward_idx = 10 + practice_pressing_num;
@@ -55,104 +38,6 @@ export const practice_stochastic_num = 10; // 10 trials practice choosing aliens
 export const practice_game_idx =
   34 + practice_pressing_num + practice_reward_num + practice_stochastic_num;
 export const practice_game_num = 20; // 20 trials to practice full game
-
-export const image_path = '../images/';
-export const image_strings = [
-  'alien1_a1.png',
-  'alien4_sp.png',
-  'tutalien1_a2.png',
-  'tutalien4_deact.png',
-  'alien1_a2.png',
-  'blackbackground.jpg',
-  'tutalien1_deact.png',
-  'tutalien4_norm.png',
-  'alien1_deact.png',
-  'earth.jpg',
-  'tutalien1_norm.png',
-  'tutalien4_sp.png',
-  'alien1_norm.png',
-  'nothing.png',
-  'tutalien1_sp.png',
-  'tutgreenplanet.jpg',
-  'alien1_sp.png',
-  'purpleplanet.jpg',
-  'alien2_a1.png',
-  'redplanet1.jpg',
-  'tutalien2_a1.png',
-  'tutrocket1_a1.png',
-  'alien2_a2.png',
-  'rocket1_a1.png',
-  'tutalien2_a2.png',
-  'tutrocket1_a2.png',
-  'alien2_deact.png',
-  'rocket1_a2.png',
-  'tutalien2_deact.png',
-  'tutrocket1_deact.png',
-  'alien2_norm.png',
-  'rocket1_deact.png',
-  'tutalien2_norm.png',
-  'tutrocket1_norm.png',
-  'alien2_sp.png',
-  'rocket1_norm.png',
-  'tutalien2_sp.png',
-  'tutrocket1_sp.png',
-  'alien3_a1.png',
-  'rocket1_sp.png',
-  'alien3_a2.png',
-  'rocket2_a1.png',
-  'tutalien3_a1.png',
-  'tutrocket2_a1.png',
-  'alien3_deact.png',
-  'rocket2_a2.png',
-  'tutalien3_a2.png',
-  'tutrocket2_a2.png',
-  'alien3_norm.png',
-  'rocket2_deact.png',
-  'tutalien3_deact.png',
-  'tutrocket2_deact.png',
-  'alien3_sp.png',
-  'rocket2_norm.png',
-  'tutalien3_norm.png',
-  'tutrocket2_norm.png',
-  'alien4_a1.png',
-  'rocket2_sp.png',
-  'tutalien3_sp.png',
-  'tutrocket2_sp.png',
-  'alien4_a2.png',
-  't.png',
-  'tutyellowplanet.jpg',
-  'alien4_deact.png',
-  'tutalien4_a1.png',
-  'alien4_norm.png',
-  'tutalien1_a1.png',
-  'tutalien4_a2.png',
-  'fish.png',
-  'tiger.png',
-  'shark.png',
-  'turtle.png',
-  'replay.png',
-];
-
-export const all_images = image_strings.map((data) => {
-  return image_path + data;
-});
-
-// Audio to preload
-export const quiz_audio = [
-  'audio/quiz/Q1_Correct.wav',
-  'audio/quiz/Q1_Incorrect.wav',
-  'audio/quiz/Q2_Correct.wav',
-  'audio/quiz/Q2_Incorrect.wav',
-  'audio/quiz/Q3_Correct.wav',
-  'audio/quiz/Q3_Incorrect.wav',
-  'audio/beep_loop.wav',
-  'audio/instructions/treasure_feedback.wav',
-  'audio/instructions/break_1.wav',
-  'audio/instructions/break_2.wav',
-  'audio/instructions/break_3.wav',
-  'audio/instructions/query.wav',
-  'audio/instructions/end_of_task.wav',
-];
 
 // Global position variables
 export const height = window.innerHeight;
@@ -200,7 +85,6 @@ export const font_size = picture_height * 25 / 758;
 /** Texts **/
 export const query_text = 'Which spaceship went mostly to the red planet?';
 export const instructions = [];
-export const audio_instructions = [];
 
 // black background
 instructions[0] = [
@@ -470,355 +354,3 @@ instructions[18] = [
   ],
 ];
 
-
-export const left_images = [];
-export const right_images = [];
-export const center_images = [];
-export const reward_images = [];
-export const audio_files = [];
-export const button_images = [];
-
-let curr_instructs;
-for (let i = 0; i < instructions.length; i += 1) {
-  curr_instructs = instructions[i];
-  left_images[i] = [];
-  right_images[i] = [];
-  reward_images[i] = [];
-  center_images[i] = [];
-  audio_files[i] = [];
-  button_images[i] = [];
-  for (let j = 0; j < curr_instructs.length; j += 1) {
-    left_images[i][j] = null;
-    right_images[i][j] = null;
-    center_images[i][j] = null;
-    reward_images[i][j] = null;
-    button_images[i][j] = 'images/button.jpeg';
-  }
-}
-
-reward_images[3][0] = reward_string;
-reward_images[3][1] = null_string;
-
-center_images[4][0] = 'images/tutalien3_norm.png';
-center_images[4][1] = 'images/tutalien3_norm.png';
-center_images[9][0] = 'images/tutalien2_norm.png';
-
-right_images[1][0] = 'images/tutrocket1_norm.png';
-left_images[1][0] = 'images/tutrocket2_norm.png';
-
-right_images[2][0] = 'images/tutalien1_norm.png';
-left_images[2][0] = 'images/tutalien2_norm.png';
-right_images[2][1] = 'images/tutalien1_norm.png';
-left_images[2][1] = 'images/tutalien2_norm.png';
-right_images[2][2] = 'images/tutalien1_norm.png';
-left_images[2][2] = 'images/tutalien2_norm.png';
-
-right_images[6][0] = 'images/tutalien1_norm.png';
-right_images[7][0] = 'images/tutalien2_norm.png';
-
-left_images[6][0] = 'images/tutalien2_norm.png';
-left_images[7][0] = 'images/tutalien1_norm.png';
-
-right_images[13][0] = 'images/tutrocket1_norm.png';
-right_images[13][1] = 'images/tutrocket1_norm.png';
-right_images[14][0] = 'images/tutrocket1_norm.png';
-right_images[14][1] = 'images/tutrocket1_norm.png';
-right_images[14][2] = 'images/tutrocket1_norm.png';
-right_images[14][3] = 'images/tutrocket1_sp.png';
-right_images[14][4] = 'images/tutrocket1_norm.png';
-right_images[14][5] = 'images/tutrocket1_norm.png';
-
-left_images[13][0] = 'images/tutrocket2_norm.png';
-left_images[13][1] = 'images/tutrocket2_norm.png';
-left_images[14][0] = 'images/tutrocket2_norm.png';
-left_images[14][1] = 'images/tutrocket2_norm.png';
-left_images[14][2] = 'images/tutrocket2_norm.png';
-left_images[14][3] = 'images/tutrocket2_sp.png';
-left_images[14][4] = 'images/tutrocket2_norm.png';
-left_images[14][5] = 'images/tutrocket2_norm.png';
-
-export const instructions_backgrounds = [
-  'images/blackbackground.jpg',
-  'images/earth.jpg',
-  'images/tutgreenplanet.jpg',
-  'images/blackbackground.jpg',
-  'images/tutyellowplanet.jpg',
-  'images/blackbackground.jpg',
-  'images/tutgreenplanet.jpg',
-  'images/tutgreenplanet.jpg',
-  'images/blackbackground.jpg',
-  'images/blackbackground.jpg',
-  'images/earth.jpg',
-  'images/tutgreenplanet.jpg',
-  'images/tutyellowplanet.jpg',
-  'images/earth.jpg',
-  'images/earth.jpg',
-  'images/blackbackground.jpg',
-];
-
-audio_instructions[0] = [['audio/instructions/I_0.1.wav']];
-audio_instructions[1] = [['audio/instructions/I_1.1.wav']];
-
-
-audio_instructions[2] = [
-  ['audio/instructions/I_2.1.wav'],
-  ['audio/instructions/I_2.2.wav'],
-  ['audio/instructions/I_2.3.wav'],
-];
-
-
-audio_instructions[3] = [
-  ['audio/instructions/I_3.1.wav'],
-  ['audio/instructions/I_3.2.wav'],
-  ['audio/instructions/I_3.3.wav'],
-  ['audio/instructions/I_3.4.wav'],
-];
-
-audio_instructions[4] = [
-  ['audio/instructions/I_4.1.wav'],
-  ['audio/instructions/I_4.2.wav'],
-];
-
-audio_instructions[5] = [
-  ['audio/instructions/I_5.1.wav'],
-];
-
-audio_instructions[6] = [
-  ['audio/instructions/I_6.1.wav'],
-];
-
-audio_instructions[7] = [
-  ['audio/instructions/I_7.1.wav'],
-];
-
-audio_instructions[8] = [
-  ['audio/instructions/I_8.1.wav'],
-  ['audio/instructions/I_8.2.wav'],
-  ['audio/instructions/I_8.3.wav'],
-];
-
-audio_instructions[9] = [
-  ['audio/instructions/I_9.1.wav'],
-  ['audio/instructions/I_9.2.wav'],
-  ['audio/instructions/I_9.3.wav'],
-  ['audio/instructions/I_9.4.wav'],
-  ['audio/instructions/I_9.5.wav'],
-  ['audio/instructions/I_9.6.wav'],
-];
-
-audio_instructions[10] = [
-  ['audio/instructions/I_10.1.wav'],
-];
-
-audio_instructions[11] = [
-  ['audio/instructions/I_11.1.wav'],
-];
-
-audio_instructions[12] = [
-  ['audio/instructions/I_12.1.wav'],
-];
-
-audio_instructions[13] = [
-  ['audio/instructions/I_13.1.wav'],
-  ['audio/instructions/I_13.2.wav'],
-];
-
-audio_instructions[14] = [
-  ['audio/instructions/I_14.1.wav'],
-  ['audio/instructions/I_14.2.wav'],
-  ['audio/instructions/I_14.3.wav'],
-  ['audio/instructions/I_14.4.wav'],
-  ['audio/instructions/I_14.5.wav'],
-  ['audio/instructions/I_14.6.wav'],
-];
-
-audio_instructions[15] = [
-  ['audio/instructions/I_15.1.wav'],
-];
-
-audio_instructions[16] = [
-  ['audio/instructions/I_16.1.wav'],
-  ['audio/instructions/I_16.2.wav'],
-  ['audio/instructions/I_16.3.wav'],
-  ['audio/instructions/I_16.4.wav'],
-];
-
-audio_instructions[17] = [
-  ['audio/instructions/I_17.1.wav'],
-];
-
-audio_instructions[18] = [
-  ['audio/instructions/I_18.1.wav'],
-];
-
-let t_i;
-let curr_page;
-let curr_side;
-
-const create_instructions = (
-    image, texts, audio_files, sect_right_texts,
-    sect_left_texts, sect_center_texts, sect_reward_texts) => {
-  'use strict';
-  const instruction_pages = [];
-  for (t_i = 0; t_i < texts.length; t_i += 1) {
-    curr_page = {
-      type: 'd3-instructions',
-      stimulus: image,
-      right_text: sect_right_texts[t_i],
-      left_text: sect_left_texts[t_i],
-      center_text: sect_center_texts[t_i],
-      reward_string: sect_reward_texts[t_i],
-      choices: jsPsych.ALL_KEYS,
-      prompt: texts[t_i],
-      audio_stimulus: audio_files[t_i],
-    };
-    instruction_pages.push(curr_page);
-  }
-  return instruction_pages;
-};
-
-// Create instructions
-export let curr_instructions = [];
-for (let i = 0; i < instructions.length; i += 1) {
-  curr_instructions = curr_instructions.concat(
-      create_instructions(
-          instructions_backgrounds[i],
-          instructions[i],
-          audio_instructions[i],
-          right_images[i],
-          left_images[i],
-          center_images[i],
-          reward_images[i],
-          button_images[i],
-      ),
-  );
-}
-
-
-// INSERT PRACTICE
-// insert 4 selection practice trials on instructions page 5
-for (let i = 0; i < (practice_pressing_num - 1); i += 1) {
-  curr_instructions.splice(practice_pressing_idx, 0, {
-    type: 'd3-animate-choice',
-    timeout: false,
-    choices: [left_key, right_key],
-    planet_text: 'images/tutgreenplanet.jpg',
-    right_text: 'images/tutalien1',
-    left_text: 'images/tutalien2',
-    prompt: ['Now try another one'],
-    trial_duration: choicetime,
-  });
-}
-curr_instructions.splice(practice_pressing_idx, 0, {
-  type: 'd3-animate-choice',
-  timeout: false,
-  choices: [left_key, right_key],
-  planet_text: 'images/tutgreenplanet.jpg',
-  right_text: 'images/tutalien1',
-  left_text: 'images/tutalien2',
-  trial_duration: choicetime,
-});
-
-// insert 10 treasure asking practice trials
-for (let i = 0; i < practice_reward_num; i += 1) {
-  curr_instructions.splice(practice_reward_idx, 0, {
-    type: 'd3-animate-choice',
-    timeout: false,
-    trial_row: reward_instructions_payoff,
-    choices: [left_key, right_key],
-    planet_text: 'images/tutyellowplanet.jpg',
-    right_text: () => {
-      if (curr_side === true) {
-        return 'images/tutalien3';
-      }
-      return null;
-    },
-    left_text: () => {
-      if (curr_side === true) {
-        return null;
-      }
-      return 'images/tutalien3';
-    },
-    trial_duration: choicetime,
-  });
-}
-
-// insert 10 asking green aliens for reward trials
-for (let i = 0; i < practice_stochastic_num; i += 1) {
-  curr_instructions.splice(practice_stochastic_idx, 0, {
-    type: 'd3-animate-choice',
-    timeout: false,
-    trial_row: instructions_payoff,
-    choices: [left_key, right_key],
-    planet_text: 'images/tutgreenplanet.jpg',
-    right_text: 'images/tutalien1',
-    left_text: 'images/tutalien2',
-    trial_duration: choicetime,
-  });
-}
-
-// remove the instructions about aliens on either side
-curr_instructions.splice(27, 3);
-
-// Insert audio test trials
-// curr_instructions.splice(1, 0, {
-//   type: 'html-keyboard-response',
-//   choices: jsPsych.ALL_KEYS,
-//   stimulus: 'Welcome! Press the space bar to begin.',
-// });
-
-// curr_instructions.splice(2, 0, {
-//   type: 'audio-keyboard-response',
-//   stimulus: 'audio/beep_loop.wav',
-//   choices: jsPsych.ALL_KEYS,
-//   prompt: 'You should now hear beeps playing. If so, ' +
-//                'press the space bar to proceed to the audio test.',
-// });
-
-
-// curr_instructions.splice(3, 0, {
-//   type: 'audio-button-response',
-//   stimulus: 'audio/fish.mp3',
-//   choices: ['repeat', 'fish', 'tiger', 'turtle', 'shark'],
-//   correct_answer: 1,
-//   prompt: 'Click on the word that you just heard.',
-//   incorrect_prompt: 'Incorrect, please adjust your volume and try again.',
-//   margin_vertical: '40px',
-//   margin_horizontal: '10px',
-//   button_html:[
-//       '<img src='images/replay.png' height='200px' width='200px'/>',
-//       '<img src='images/fish.png' height='200px' width='200px'/>',
-//       '<img src='images/tiger.png' height='200px' width='200px'/>',
-//       '<img src='images/turtle.png' height='200px' width='200px'/>',
-//       '<img src='images/shark.png' height='200px' width='200px'/>'
-//   ],
-//   post_trial_gap: 1000
-// });
-
-
-// curr_instructions.splice(4, 0, {
-//   type: 'audio-button-response',
-//   stimulus: 'audio/tiger.mp3',
-//   choices: ['repeat', 'turtle', 'shark', 'fish', 'tiger'],
-//   correct_answer: 4,
-//   prompt: 'Again, click on the word that you just heard.',
-//   incorrect_prompt: 'Incorrect, please adjust your volume and try again.',
-//   margin_vertical: '40px',
-//   margin_horizontal: '10px',
-//   button_html:[
-//       '<img src='images/replay.png' height='200px' width='200px'/>',
-//       '<img src='images/turtle.png' height='200px' width='200px'/>',
-//       '<img src='images/shark.png' height='200px' width='200px'/>',
-//       '<img src='images/fish.png' height='200px' width='200px'/>',
-//       '<img src='images/tiger.png' height='200px' width='200px'/>'
-//   ],
-//   post_trial_gap: 1000
-// });
-
-// add full screen after audio test
-// curr_instructions.splice(5, 0, {
-//   type: 'fullscreen',
-//   message: '<p> Press start to enter ' +
-//              'full-screen mode and start the game. </p>',
-//   button_label: 'Start',
-//   fullscreen_mode: true});
