+++ src/lib/plugins/two-step-explicit-choice.js
@@ -8,6 +8,8 @@
  * KN 5/23/20
  */
 import {select, interval} from 'd3';
+import {experiment} from '../..';
+import consola from 'consola';
 import {
   width,
   height,
@@ -23,7 +25,6 @@ import {
   text_start_x,
   instructions_text_start_y,
   font_size,
-  reward_string,
   reward_x,
   reward_y,
   reward_size,
@@ -120,6 +121,8 @@ jsPsych.plugins['two-step-explicit-choice'] = (() => {
   };
 
   plugin.trial = function(display_element, trial) {
+    consola.debug(`Running trial:`, trial.type);
+
     const new_html = `<div id='container' class='exp-container'></div>`;
     display_element.innerHTML = new_html;
 
@@ -467,9 +470,9 @@ jsPsych.plugins['two-step-explicit-choice'] = (() => {
       }
 
       if (reward) {
-        return reward_string;
+        return experiment.getStimuli().getCollection()['t.png'];
       } else {
-        return null_string;
+        return experiment.getStimuli().getCollection()['nothing.png'];
       }
     }
   };
